SimtesOPF.get_gen_data

SimtesOPF.get_load_data

SimtesOPF.get_spot_prices_data

SimtesOPF.get_gen_data — Function

get_gen_data(grid_data_path::String,
             n_days::Int=30,
             gen_index::Union{Nothing, Int}=nothing)

Seleciona um perfil e lê o perfil de geração referente ao perfil selecionado.

Input

grid_data_path – indica o diretório onde estão os dados de geração.

n_days – (optional, default: 30) Número de dias de simulação

gen_index – (optional, default: nothing) caso um valor seja passado o perfil selecionado será daquele indice, caso contrario será selecionado um pefil aleatório.

Output

Dataframe com o perfil de geração ativa e reativa.

SimtesOPF.get_load_data — Function

get_load_data(grid_data_path::String,
              n_days::Int=30
              load_index::Union{Nothing, Int}=nothing)

Seleciona um perfil e lê o perfil de carga referente ao perfil selecionado.

Input

grid_data_path – (griddatapath::String) indica o diretório onde estão os dados de carga.

n_days – (optional, default: 30) Número de dias de simulação

load_index – (optional, default: nothing) caso um valor seja passado o perfil selecionado será daquele indice, caso contrario será selecionado um pefil aleatório.

Output

Dataframe com o perfil de carga ativa e reativa. SALVE NETO

SimtesOPF.get_spot_prices_data — Function

get_spot_prices_data(grid_data_path::String,
                     n_days::Int=30,
                     spot_index::Union{Nothing, Int}=nothing)

Le dados referentes ao preço spot de energia.

Input

grid_data_path – indica o diretório onde estão os dados de carga.

n_days – (optional, default: 30) Número de dias de simulação

spot_index – (optional, default: nothing) caso um valor seja passado o perfil selecionado será daquele indice, caso contrario será selecionado indice aleatorio.

Output

Dataframe com o perfil de carga ativa e reativa.